fix: Moving Table to a new file component
