+++ package.json
@@ -9,7 +9,8 @@
 		"lint": "eslint .",
 		"fixlint": "eslint --fix .",
 		"test": "vitest",
-		"coverage": "vitest run --coverage"
+		"coverage": "vitest run --coverage",
+		"test:ui": "vitest --ui"
 	},
 	"dependencies": {
 		"@ant-design/icons": "^4.7.0",
