+++ src/components/index.jsx
@@ -1,4 +1,5 @@
 import Breadcrumbs from './Breadcrumbs'
 import Searchbar from './Searchbar'
+import TableData from './TableData'
 
-export { Breadcrumbs, Searchbar }
+export { Breadcrumbs, Searchbar, TableData }
